Add support for Django 3.0
This primarily changes the use of deprecated imports, and adds
some logic for retaining support for python2.7 imports where needed.

Note that JSONField and multiselect, which are 3rd party modules,
do not yet support django 3 and both need updating for this to work:
JSONField needs the changes in imports, and multiselect needs a
method signature to be updated. I'll send pull requests to those
maintainers.